wire editor
